fix charts ui state
