+++ client_restaurant/shared/src/commonMain/kotlin/di/ScreenModelModule.kt
@@ -14,7 +14,7 @@ import presentation.restaurantSelection.RestaurantSelectionScreenModel
 
 val screenModelModule = module {
     factoryOf(::LoginScreenModel)
-    factory { (restaurantId: String) -> MainScreenModel(restaurantId, get(),get(),get()) }
+    factory { (restaurantId: String) -> MainScreenModel(restaurantId, get(),get()) }
     factoryOf(::IMealCreationScreenModel)
     factoryOf(::OrderScreenModel)
     factoryOf(::RestaurantInformationScreenModel)
