--- client_restaurant/shared/src/commonMain/kotlin/domain/usecase/GetBarChartDataUseCase.kt
@@ -1,15 +1,0 @@
-package domain.usecase
-
-import domain.entity.BarsParameters
-import domain.gateway.IFakeRemoteGateway
-
-interface IGetBarChartDataUseCase {
-    suspend fun getBarChartData(): BarsParameters
-}
-class GetBarChartDataUseCase(private val fakeRemoteGateWay: IFakeRemoteGateway)
-    : IGetBarChartDataUseCase {
-    override suspend fun getBarChartData(): BarsParameters {
-        return fakeRemoteGateWay.getBarChartDate()
-    }
-
-}
